refactor: move uni command position
